# Speed up baselineGrid.py and cut prints