In the original form, the `geo2rdr` method of Orbit.py took a couple hundred milliseconds to run, and it also printed `Polynomial Order: 0 - by - 0` each time.
For my stripmap stack run, this would print out ~20k times per SLC.

```python
In [60]: %time orb.geo2rdr(target)
Polynomial Order: 0 - by - 0
0
CPU times: user 181 ms, sys: 11.8 ms, total: 193 ms
Wall time: 192 ms
```
It also meant that each `baselineGrid.py` run would take over an hour, and the entire step would take about a day for my stack of 25.

I had initially just changed the `run_07_grid_baseline` to kick off a bunch of them in parallel, since they are only using one CPU core to run. But I noticed two quick fixes for the python `geo2rdr`:

1. Each time, a `Poly2D` object is created, because we were calling `sOrb.geo2rdr(target)`, and `geo2rdr` had this
```python
        if doppler is None:
            doppler = Poly2D()
            doppler.initPoly(azimuthOrder=0, rangeOrder=0, coeffs=[[0.]])
            wvl = 0.
```


Creating that once and passing it through saved ~20% on it, but it also cuts out the 1000s of "Polynomial Order 0" prints
```python
In [65]: %time orb.geo2rdr(target, doppler=doppler, wvl=0)
CPU times: user 164 ms, sys: 4.99 ms, total: 169 ms
Wall time: 168 ms
```

2. adding a check for convergence in the loop at the bottom, like the fortran version has https://github.com/isce-framework/isce2/blob/main/components/zerodop/geo2rdr/src/geo2rdr.f90#L300-L304 

```python
    # Start the previous guess tracking with dummy value
    t_prev_guess = tguess + datetime.timedelta(seconds=10)
...
            if abs(tguess - t_prev_guess).total_seconds() < 5e-9:
                break
            else:
                t_prev_guess = tguess
```
makes it break after a few iterations (for the testing i added a print with the total iterations):
```python
In [76]: %time orb.geo2rdr(target, doppler=doppler, wvl=0)
Total iters: 3
CPU times: user 11.1 ms, sys: 1.02 ms, total: 12.1 ms
Wall time: 11.3 ms
```